# Optimize IN predicate