--- a/velox/exec/tests/PlanBuilder.cpp
+++ b/velox/exec/tests/PlanBuilder.cpp
@@ -37,9 +37,10 @@ std::vector<std::string> makeNames(const std::string& prefix, int size) {
 
 std::shared_ptr<const core::ITypedExpr> parseExpr(
     const std::string& text,
-    std::shared_ptr<const RowType> rowType) {
+    std::shared_ptr<const RowType> rowType,
+    memory::MemoryPool* pool) {
   auto untyped = parse::parseExpr(text);
-  return core::Expressions::inferTypes(untyped, rowType, nullptr);
+  return core::Expressions::inferTypes(untyped, rowType, pool);
 }
 } // namespace
 
@@ -114,7 +115,8 @@ PlanBuilder& PlanBuilder::project(
   }
   std::vector<std::shared_ptr<const core::ITypedExpr>> expressions;
   for (auto& projection : projections) {
-    expressions.push_back(parseExpr(projection, planNode_->outputType()));
+    expressions.push_back(
+        parseExpr(projection, planNode_->outputType(), pool_));
   }
   planNode_ = std::make_shared<core::ProjectNode>(
       nextPlanNodeId(),
@@ -126,7 +128,9 @@ PlanBuilder& PlanBuilder::project(
 
 PlanBuilder& PlanBuilder::filter(const std::string& filter) {
   planNode_ = std::make_shared<core::FilterNode>(
-      nextPlanNodeId(), parseExpr(filter, planNode_->outputType()), planNode_);
+      nextPlanNodeId(),
+      parseExpr(filter, planNode_->outputType(), pool_),
+      planNode_);
   return *this;
 }
 
@@ -228,7 +232,7 @@ PlanBuilder& PlanBuilder::aggregation(
     }
 
     auto expr = std::dynamic_pointer_cast<const core::CallTypedExpr>(
-        parseExpr(agg, planNode_->outputType()));
+        parseExpr(agg, planNode_->outputType(), pool_));
     aggregateExprs.emplace_back(expr);
   }
 
@@ -448,7 +452,7 @@ PlanBuilder& PlanBuilder::hashJoin(
   auto resultType = concat(leftType, rightType);
   std::shared_ptr<const core::ITypedExpr> filterExpr;
   if (!filterText.empty()) {
-    filterExpr = parseExpr(filterText, resultType);
+    filterExpr = parseExpr(filterText, resultType, pool_);
   }
   auto outputType = extract(resultType, output);
   auto leftKeyFields = fields(leftType, leftKeys);