--- a/velox/exec/tests/PlanBuilder.h
+++ b/velox/exec/tests/PlanBuilder.h
@@ -17,12 +17,17 @@
 #include <velox/core/Expressions.h>
 #include <velox/core/ITypedExpr.h>
 #include <velox/core/PlanNode.h>
+#include "velox/common/memory/Memory.h"
 
 namespace facebook::velox::exec::test {
 
 class PlanBuilder {
  public:
-  explicit PlanBuilder(int planNodeId = 0) : planNodeId_{planNodeId} {}
+  PlanBuilder(int planNodeId = 0, memory::MemoryPool* pool = nullptr)
+      : planNodeId_{planNodeId}, pool_{pool} {}
+
+  explicit PlanBuilder(memory::MemoryPool* pool)
+      : planNodeId_{0}, pool_{pool} {}
 
   PlanBuilder& tableScan(const RowTypePtr& outputType);
 
@@ -219,5 +224,6 @@ class PlanBuilder {
 
   int planNodeId_;
   std::shared_ptr<core::PlanNode> planNode_;
+  memory::MemoryPool* pool_;
 };
 } // namespace facebook::velox::exec::test